renaming memory func to pgr_palloc

+++ tools/template/src/function1.c
@@ -74,7 +74,7 @@ process( char* edges_sql,
 
     PGR_DBG("Load data");
     pgr_edge_t *edges = NULL;
-    int64_t total_tuples = 0;
+    size_t total_tuples = 0;
     pgr_get_data_5_columns(edges_sql, &edges, &total_tuples);
 
     if (total_tuples == 0) {
@@ -115,9 +115,9 @@ PGDLLEXPORT Datum
 #endif
 MY_FUNCTION_NAME(PG_FUNCTION_ARGS) {
     FuncCallContext     *funcctx;
-    size_t              call_cntr;
-    size_t               max_calls;
-    TupleDesc            tuple_desc;
+    uint32_t            call_cntr;
+    uint32_t            max_calls;
+    TupleDesc           tuple_desc;
 
     /**************************************************************************/
     /*                          MODIFY AS NEEDED                              */
@@ -159,7 +159,7 @@ MY_FUNCTION_NAME(PG_FUNCTION_ARGS) {
         /*                                                                             */
         /*******************************************************************************/
 
-        funcctx->max_calls = result_count;
+        funcctx->max_calls = (uint32_t) result_count;
         funcctx->user_fctx = result_tuples;
         if (get_call_result_type(fcinfo, NULL, &tuple_desc) != TYPEFUNC_COMPOSITE)
             ereport(ERROR,
